+++ src/assets/icons/angle-right-icon/angle-right-icon.tsx
@@ -1,13 +1,20 @@
-import { Component, h } from '@stencil/core';
+import { Component, h, Prop } from '@stencil/core';
 
 @Component({
   tag: 'mvx-angle-right-icon',
+  styleUrl: 'angle-right-icon.scss',
   shadow: true,
 })
 export class AngleRightIcon {
+  @Prop() class?: string;
+
   render() {
     return (
-      <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" width="10" height="16">
+      <svg
+        viewBox="0 0 320 512"
+        xmlns="http://www.w3.org/2000/svg"
+        class={{ 'angle-right-icon': true, [this.class]: Boolean(this.class) }}
+      >
         <path d="M278.6 233.4c12.5 12.5 12.5 32.8 0 45.3l-160 160c-12.5 12.5-32.8 12.5-45.3 0s-12.5-32.8 0-45.3L210.7 256 73.4 118.6c-12.5-12.5-12.5-32.8 0-45.3s32.8-12.5 45.3 0l160 160z" />
       </svg>
     );