+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- (Disable sourceMap files)[https://github.com/multiversx/mx-sdk-dapp-ui/pull/113]
+
 ## [[0.0.4](https://github.com/multiversx/mx-sdk-dapp-ui/pull/112)] - 2025-05-26
 
 - [Clean up unlock panel](https://github.com/multiversx/mx-sdk-dapp-ui/pull/111)